# endpoints/src/main/java/fr/epita/assistants/presentation/rest/request/ReverseRequest.java
package fr.epita.assistants.presentation.rest.request;

public class ReverseRequest {
    private String content;

    public ReverseRequest(String content) {
        this.content = content;
    }

    public String getContent() {
        return this.content;
    }
}